rework flow of do_raw_disk_image_create_partitions
This is mostly done to align the different device type and how the
initial raw disk image is processed. The "beaglebone" type differed
a lot from how it is done for "raspberrypi" and "qemu", the big
difference was that it did not extract the necessary images from
the raw disk image. move the them to the mender disk image where
addition processing was done. Instead it processed directly on
the raw disk image.

To be able to optmize the workflows and stages, they have to be
executed in a similar fashion on all the device types.

Now do_raw_disk_image_create_partitions will always extract the
the images from the raw disk image, instead of doing it in the
device specific do_make_sdimg_* functions which contained a lot
of duplication.

By aligning the workflow we have can also get rid of the
beagblebone specific stage-3, and it can use the "generic" one.

Changelog: Optimizations to speed up conversion, utilizing sparse
images.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>