remove expert commands
Note that there is not much actual code removed, but this commit
only removes the exposure to the user.

There is a big re-write incoming and is it not feasable to maintain
"convenience" commands that are exposed to the users. The expert
command paths are not tested in any automated fashion and the
likely hood of them breaking is pretty high.

We can always bring commands if there are requests from users.

Note that that the "raw-disk-image-shrink-rootfs" expert command is
promoted to a "General command" and is not removed. This has shown to
provide value to our users.

Changelog: Title

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>