always run 'image_shrink_rootfs' in 'do_from_raw_disk_image'
This message has lately generated some confusion,

    *** Data partition size set to default value: 128MB ***
    1/9 Repartitioning raw disk image...
     Detected raw disk image with 2 partition(s).
     Calculating partitions' sizes of the Mender image.
     Adjust Mender disk image size to the total storage size (2048MB).
     Defined total storage size of MB is too small.
     Minimal required storage is 60056MB. Aborting.

There are two things here that are hard to understand:

- Why does it not respect the 'storage-total-size-mb' setting of 2048MB?
- Where does the 60GB requirement come frome?

This normally happens if you have booted an image once on e.g Raspbian,
which will expand the rootfs to occupy the reamining blocks of the SD
card. And if you then do a "dd" dump of that image you will get that
expanded rootfs size. We normally recommend to run
"raw-disk-image-shrink-rootfs" on these type of images before
"from-raw-disk-image" command, which will ensure that the rootfs is
reduced in size.

To remedy the confusion here, simply run the
"raw-disk-image-shrink-rootfs" unconditionally when running
"from-raw-disk-image", this way it will not matter if the input image
has been "expanded" and the '--storage-total-size-mb' will always be
respected assuming that that the actual rootfs size does not overflow
the '--storage-total-size-mb' setting.

Because 'raw-disk-image-shrink-rootfs' is now a default, we also remove
the external command for this.

Changelog: shrink expanded rootfs when running "from_raw_disk_image"

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>
(cherry picked from commit c4c3c9ec95dc4406d4592c5b797237b1066d3d5f)