do not fail conversion in disk_image_shrink_rootfs
That is if it fails, proceed with conversion without it.

The current logic only supports shrinking a rootfs on BBB and RPi
devices (this is deducted from $raw_disk_counts) and calling
'disk_image_shrink_rootfs' would fail on 'qemux86_64' device.

Shrinking the rootfs is not a hard-requirement and is only an
optimization (mainly for RPi devices).

Changelog: None

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>