rpi: init_resize: pull changes done to cmdline.txt from upstream script
It seems some additional manipulation was added to cmdline.txt in the upstream
init_resize.sh script, the part that removed the 'quiet' kernel argument is what
will add the Linux kernel console on HDMI again and if 'quiet' is present nothing is
shown on the HDMI output resulting in a "black screen".

MEN-2673

Changelog: Fix console not being available on HDMI on Raspberry Pi in Raspbian Buster.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>