Fix GitLab Docker build using default docker-build script
So there are other arguments that need to be passed to docker build,
namely toolchain related, therefore it is better to use the default
script and just override the image name.

Also, removing the Mender version from the GitLab pipeline (and hence
using the one defined in build-docker) removes one point of
inconsistency.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>