Update the Readme file
Replace the explicit image name with a variable

The date of the raspbian buster images were wrong, as the downloaded image had a
date in june, and the convert script, had a date of march, thus the command did
not work out of the box.

This fix moves the script to use a variable for the images, to mitigate these
issues for the future.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>

Make the sections in the Readme more explicitly separate

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>

Remove passive aggressive docker comment in Readme

Unecessary

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>

Fill the first paragraphs to 80 columns, to unify the style

Some parts of the Readme wrapped at 80, and other did not.

Unify.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>

Make all shell sections in the readme use bash formatting

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>