Don't use --apparent-size to calculate rootfs total size.
Not using this option produces a more realistic result, because block
padding has to be taken into account when calculating the totals. It
is not a perfect method however, since the calculation becomes
dependent on the block size of the filesystem on the build machine.
4096 is a *very* common size though, and almost certainly will be the
size used both on the build host and in the resulting filesystem.

Changelog: Changed the calculation of occupied space on the rootfs
partition to a more accurate formula.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>