MEN-2794: Adjust the expected S3 locations for the binaries
The S3 path now follows the pattern:
/mender/mender-convert/uboot/<dev-family>/<dev-config>-<uboot-ver>.tar

For RPi and BBB these are automatically published form the CI Pipeline
of repo mender-convert-integration-scripts while for RockPro64 they have
been manually moved (as we expect this to be a one time integration).

Reworked a bit also the *_BINARIES variable so that we can keep the
u-boot version in the base config files and let the childs only define
the a config name.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>