Make sure that DTB files are included on boot partition.
This is important so that the bootloader (U-Boot) can use the DTB
files when chainloading GRUB via UEFI. The logic for doing this is
already present in `distro_bootcmd`, but because we were not placing
the DTB files in the correct folder, this part was skipped.

Changelog: Fix certain kernels hanging on boot. In particular, recent
versions of Debian for Beaglebone was affected, but several other
boards using UEFI may also have been affected.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>