Revert Beaglebone test image to 2018-08-30 revision.
This image is old enough to have a U-Boot version from before the boot
problems in MEN-2404 were introduced (>= v2018.09). At the time of
writing, no image has been released which has a U-Boot from after the
problem was fixed (> v2019.07), but we expect it will eventually be
released, and then we can bump to that version.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>