Report which error occured before exiting
This adds more accurate descriptions of what error actually occured during
script execution, as compared to previously, where it would silently fail, and
it could be hard to figure out exactly where the error occured.

Now, upon an error, the script will always output these lines (or similar) at the end:

```bash
2020-04-23 18:53:34 [ERROR] [mender-convert] mender-convert failed
2020-04-23 18:53:34 [DEBUG] [mender-convert-extract] When running: (./mender-convert-extract:109): run_and_log_cmd():

	dd if=/dev/zer of=work/boot-generated.vfat count=40 bs=1M status=none
	dd: failed to open '/dev/zer': No such file or directory

2020-04-23 18:53:34 [ERROR] [mender-convert] mender-convert exit code: 1
```

Which is a description of which part of the program failed. What the exit code
was, and the last command entry in the debug log, and if any, the collected
output from the command is printed.

Changelog: Print improved error diagnostics before exiting on an error

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>