test(automatic-decompression): Add tests for automatic decompression
This change required some changes to the current test scripts. Most notably, the
whole conversion is now run two times. One for the compressed input image, and
one for the already decompressed image.

Also, the test framework was cleaned up significantly, and the number of
arguments passed to the 'convert_and_test' function has been decreased to remove
the redundancies that were present (like passing the image name, and the image
name with the extension, in two separate parameters).

These changes are a result of the changes introduced with:

* MEN-3052: https://tracker.mender.io/browse/MEN-3052

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>