[tests, CI, dependabot] Remove --no-pull, update tests, add dependabot
Remove all the logic for the special handling of this submodule, and
treat it instead a regular Git submodule.

This commit updates the submodule to latest, removes the --no-pull
functionality, and removes the ignore of the submodule to make
developers aware

To ease maintenance, add also dependabot to pay attention to the future
updates of the submodule.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>
(cherry picked from commit 721b9f3ce2c55998fee62dcaf2e5cb3178a73661)