MEN-4256 Do not install the client by default on Prebuilt RPi images
This adds the:

* `MENDER_CLIENT_INSTALL` option, to allow an image to come without the
mender-client installed by default.

The option is on by default, but turned off for the custom image configuration
we use for building the images we provide as a part of our documentation onboarding.

Changelog: Added an option `MENDER_CLIENT_INSTALL=y/n`, in order to
configure the installation of the Mender client into the converted image.
Defaults to `yes`.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>
Signed-off-by: Lluis Campos <lluis.campos@northern.tech>