MEN-4462: Add a demo configuration config script
This moves the device configuration to a config script in configs/, from the
previous setup in the bootstrap scripts.

It does so by creating a new resources/ folder in the top-level directory, which
is then populated with the configuration file, and the server parameters only.

When the conversion is run with the demo configuration
configs/mender_convert_demo_config the conversion will install all the add-on's
available, and modify the configuration file in resources to align with our
regular Mender demo configuration.

Changelog: Add a demo configuration in configs/mender_convert_demo_config which
when added at run time creates a Mender demo setup in the converted image. This
includes short polling intervals, and all add-on's installed by default.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>