Modify `ln` calls for data store links to support successive runs
Fixes MEN-5078

According to `ln` man page, the correct way to create a link on a
directory is with `ln [OPTION]... TARGET... DIRECTORY`.

Create first `DIRECTORY`, and then use the above form for the link. This
way a successive run of `ln ...` will not end up creating a
`.../mender/mender` kind of link.

Changelog: Title

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>