Fix installation of user specified versions for mender-configure add-on
It has been broken since the introduction of mender-configure.

This add-on is architecture independent, so when downloading the .deb
directly from the pool we need to look for "all" suffix instead of the
target architecture. This worked fine for "latest" and "master" versions
because the filename for these were resolved via repository's Packages.

Changelog: title

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>