MEN-5221: Keep Grub if EFI shim is found in the boot partition
In the case a boot shim is found on the boot partition, we do keep the boot
partition pristine, and only install the generated mender `grub.cfg` file we need.

In all other cases, we fall back to the old functionality of installing
mender-grub and nuking the existing bootloader.

Changelog: commit
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>