Switch to detect `grub*.efi` instead of `shim*.efi`.
The presence of the shim depends on whether Secure Boot is enabled or
not, plus the configuration of the distro. GRUB itself however, will
always be present if the distro uses GRUB with UEFI at all. So check
for that instead.

Of course, without the shim, Secure Boot will not work out of the box,
but this is a misconfigured platform, not a problem with Mender.

Cancel-Changelog: 2b805e29dcfbca1bcf9ea6339ef7755025737e56

Changelog: If `grub*.efi` preexists on the EFI partition, keep it
instead of installing our own. In all other cases, we fall back to the
old functionality of installing mender-grub and nuking the existing
bootloader.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>