test: Ignore some grub.cfg differences in `00_header` section.
Ubuntu 22.04 Jammy can trigger differences in searching for the root
device when running live as opposed to mender-convert's jailroot.
Since we have grub scripts which later set root back to the value it
should be, it is safe to ignore these differences. However, since it
is dangerous to ignore these differences *outside* `00_header`, make
the sed script fairly specific so that it is only ignored inside this
section.

Ticket: QA-554

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>