refac: Add `.git` to `.dockerignore` to enable better cache usage.
Because the inside of the Docker container will then no longer have
access to the Git repository, we need to move all Git operations
outside the container. Only version detection is affected, and Git is
already required on the running host to detect the image version to
use, so this should not be a problem.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit df73e13941a30cdf80465801a010db827d560c0b)