fix: Remove unneeded and potentially harmful storage assignment.
Defining the `mender_grub_storage_device` variable manually has
actually not been needed in grub-mender-grubenv for a long time,
because it figures it out on its own using the `root` variable, which
GRUB fills in during boot. This works better when storage devices
change, like when inserting a USB stick, for example.

Although it is definitely unneeded, it's not entirely clear why it
messes with the boot in certain instances, since this assignment
should happen before any assignment from `root`. But a user has
confirmed that this helped.

Changelog: Remove unneeded `mender_grub_storage_device` assignment
which could boot from the wrong device if a USB stick was inserted.

Ticket: CE-321

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>