feat: Switch to install packages using Debian package manager.
This requires us to move the package installations further down in the
modify script, because some package scripts depend on whether or not
`/var/lib/mender` has been set up as a symlink or not. So make sure
that happens first.

Changelog: Switch to install packages using Debian package manager.
This means that the files are no longer simply extracted into place,
which has been a big source of conflict with the package manager when
doing later upgrades. It also means that a default `mender.conf` file
will be provided if `mender-setup` is installed, since it installs it
by default.

Ticket: MEN-6935

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>