fix: Fix boot problems on RaspiOS images.
There is a 64-bit kernel present in these images, which will be
preferred by the on-board bootloader, even if userspace is 32-bit.
Since the 64-bit kernel will work with 32-bit userspace just fine,
it's possible that this has gone unnoticed in the Raspberry Pi
community.

However, we are providing only a 32-bit U-Boot binary, and hence we
need to force this to load instead of the 64-bit kernel.

Changelog: Title
Ticket: MEN-7143

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>