feat: build mender-artifact from source
The mender-artifact binary which is provided as a download and used
for the mender-convert docker build so far is tied to the amd64
architecture. By building from source, the resulting binary matches
the architecture of the built container.
As a side effect, the dependency on the binary being pre-built is
removed, so the latest version of a branch can be used.

Changelog: Title
Ticket: None

Signed-off-by: TheYoctoJester <jester@theyoctojester.info>