feat: support assets directory for local data
Custom mender-convert configurations might require
data assets which are not parts of a rootfilesystem
overlay structure, such as a modified bootloader.

As an alternative to downloading them as part of the
conversion script, the /assets directory can be used to
hold and provide such assests. The directory is empty
by default in the upstream repositoy and mounted into
the docker container invoked by docker-mender-convert.

Changelog: Title
Ticket: None

Signed-off-by: Josef Holzmayr <jester@theyoctojester.info>