feat: extend Raspberry Pi support to bookworm
The Debian release 12 "bookworm" introduced the /boot/firmware
mount point for the boot partition, which is a breaking change.

The upgrade therefore involves these adjustments:
- use the "check_image_os" function to check the Linux distribution
  release requirement on the supplied image
- set MENDER_BOOT_PART_MOUNTPOINT to /boot/firmware
- remove the /uboot directory creation and related modifications
  of scripts shipped with the supplied image.
- copy the device tree overlays from the boot partition to /boot
  so they can be accessed from u-boot

To facilitate maintenance, some functionality has been factored out
into the raspberrypi_common_config file.

Currently checks for the required Linux distribution family and version,
e.g. "debian" and "12" (reflects "bookworm") are included. This is
common for both the already existing 32bit and upcoming 64bit configs,
so it prepares the version bump and aarch64 extension.

On Debian 11 "bullseye" (or derivatives) and earlier, the boot
partition is mounted to /uboot in the u-boot integration path.
As several scripts rely on the files located on the patition,
we need to adjust those.

Changelog: Commit
Ticket: None

Signed-off-by: Josef Holzmayr <jester@theyoctojester.info>