feat: Multi-platform build for `mender-convert`
This commit changes several things with the end goal of making the tool
multi-platform, namely easing the use of it on Apple M1 machines.

In short:
* Update base image to latest Ubuntu and update a dependency
* Modify Dockerfile to build platform aware images with buildx
* Switch pxz build image to Debian for easier cross-compile
* Use specific version of `mender-artifact` instead of master. The
  motivation of this specific change is to make easier the install of
  pre-built binares via deb package.
* Replace custom build/publish jobs for CI/CD with the ones from our
  templates, which are multi-platform ready
* Demote the `docker-build` script for local development only.

Changelog: `mender-convert` Docker build is now architecture aware and
it will install the native dependencies for the target architecture.
This change makes the tool compatible with Apple arm machines. The Docker
images will be published for `linux/amd64` and `linux/arm64` platforms.

Ticket: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>