feat: raspberrypi_common: support local assets
Providing assets locally can be helpful during development,
as well as for supporting customized hardware variants.
This adds support to the Raspberry Pi base configurations
for checking the assets path first if a the bootloader binaries
are provided there, and only downloads them if not present there.

Changelog: Title
Ticket: None

Signed-off-by: TheYoctoJester <jester@theyoctojester.info>