fix: Introduce a `/boot/firmware` symlink to please RPi OS tools.
It is unknown exactly what effect this will have, but it is a
requirement for at least one later fix, namely the integration with
the `init=` RPi hack. The script that is run through it, requires
access to the `/boot/firmware` directory to work correctly.

And in general it feels safer to have this link since RPi tools expect
it to be there.

Changelog: None
Ticket: MEN-5826

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>