feat(uefi): Implement support for RPi4 `cmdline.txt` boot file.
This is required to run the "firstboot" Raspberry Pi setup
correctly. It also makes sure that custom boot options in that file
are actually respected.

Changelog: Implement support for RPi4 `cmdline.txt` boot file when
booting through UEFI. Note that when changing the file it is required
to run `update-grub` as root afterwards. This will usually happen
automatically as part of package upgrades.

Ticket: MEN-5826

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>