feat: create sparse images
Instead of allocating the whole image on the disk, only allocate the parts that actually contain data. This reduces the required disk space to store or create images.
This feature is disabled by default and can be enabled with the `MENDER_SPARSE_IMAGE` configuration flag.

Changelog: Support for sparse images to save storage space. It is disabled by default and can be enabled with `MENDER_SPARSE_IMAGE` configuration flag
Ticket: None
Signed-off-by: Yannis Gerlach <y.gerlach@nexat.de>