feat: Upgrade Raspberry Pi 3 pre-built image to Debian 12 64bits
This commit restores the Raspberry Pi 3 pre-built image build and
publish, but modifying it from being based in the 32bits one to the 64
bits one.

As the commit that removed it completely did not have an explicit
changelog entry on this, the changelog here is only about "upgrading"
the image.

Changelog: Title
Ticket: MEN-7759

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>