fix: Expand the data partition into the empty space on boot
When MENDER_DATA_PART_GROWFS is "y" (which is default), x-systemd.growfs
will be set for the data partition in /etc/fstab. This option is understood
by systemd and will grow the file system to occupy the full block device.
However, the partition itself will not necessarily be expanded. For
Raspberry Pi a systemd script could be used that would run "parted"
to expand the partition. This is now done for all systems.

Ticket: MEN-8191
Changelog: Commit

Signed-off-by: John Olav Lund <john.olav.lund@northern.tech>