fix: systemd_common_config: add --fix option to parted call
In case when the GPT header does not include the full disk size (can
happen to generate smaller images), parted is going to fail because of
an exception.

Add the --fix option to fix those exceptions and allow parted to resize
the data partition.

Ticket: None
Changelog: add `--fix` option to grow_data_partition's parted call to work around exceptions.

Signed-off-by: Valentin Longchamp <valentin@longchamp.me>