fix: Add missing include to Raspberry Pi UEFI integration
While it works either way, the MENDER_COPY_BOOT_GAP should be set to no
because the upstream images don't have this gap.

Instead of just setting this config we include the whole common which
should be harmless for UEFI integration and eases maintenance of the two
integrations (uboot and UEFI).

Ticket: None
Changelog: Title

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>