chore: fix backup and restore of CA certificates
(Commit marked as chore and not fix because the bug is not released).

Bug introduced at:
* https://github.com/mendersoftware/mender-convert/commit/1f590014d7828dc43b3a975fe7a6aa56ca25ee15

The previous implementation had the flaw that `rm ... ca-certificates*`
would remove the .orig directory before it could be restored. Also it
was treating differently the certs in /etc/ssl which were backup one by
one and the ones from /etc/ca-certificates which would be copy at the
directory level.

The new code has explicit handling of the conf first and then explicit
handling of the directories, which makes it more robust.

Ticket: MEN-8783

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>