chore: Remove unmaintained GH action
Long time ago we had an automation to check for new images, for example
a new release of Raspberry Pi OS, and trigger a Pull Request updating
it.

The last successful run of this automation is from 2022. And since some
time ago it also just fails, making the reports of GitHub actions “red”.

Remove the GH Action for now and follow-up on its dedicated ticket.

Ticket: QA-1239

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>