test: decouple rpi os versions for uboot and UEFI
Move the definition of RASPIOS_IMAGE_URL to an argument to run-tests.sh
and pass it from .gitlab-ci.yaml in the jobs that need it. The
motivation behind this is to decouple to RPi OS versions for U-Boot and
UEFI. This way we can define the image in a CI variable rather than
hardcoding it in the script.

Ticket: MEN-8942

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>