ci: rename and merge the RPi OS prefix variables
Since we've separates the RPi OS image for UEFI and U-Boot, we should rename
`RASPBERRY_PI_OS_PREFIX` to `RASPBERRY_PI_OS_UBOOT_PREFIX` to make it more explicit
which version we're talking about.

Defined the variable as a top-level configuration option to avoid the
unnecessary duplication by defining the same value in two separate
variables.

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>