chore: rename and merge the RPi OS name variables
Since we've separates the RPi OS image for UEFI and U-Boot, we should rename
`RASPBIAN_NAME_CONVERT` to `RASPBERRY_PI_OS_UBOOT_NAME_CONVERT` to make it more explicit
which version we're talking about.

Defined the variable as a top-level configuration option to avoid the
unnecessary duplication by defining the same value in two separate
variables.

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>