feat: select uboot binary with boot delay disabled for rpi5
Added an override in `raspberrypi5_64_config` to use a U-Boot binary with
boot delay set to -2. This is a workaround for an issue where RPi5 with U-Boot
is stuck waiting for uart. By setting boot delay to -2, it will autoboot with
no delay and not check for abort.

Ticket: MEN-8901

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>